Move ignored file creation to own file
