+++ Sources/NextcloudFileProviderKit/Item/Item+Create.swift
@@ -475,55 +475,15 @@ extension Item {
 
         let relativePath = parentItemRelativePath + "/" + itemTemplate.filename
         guard ignoredFiles == nil || ignoredFiles?.isExcluded(relativePath) == false else {
-            Self.logger.info(
-                """
-                File \(itemTemplate.filename, privacy: .public) is in the ignore list.
-                    (Relative path: \(relativePath, privacy: .public))
-                    Will not propagate creation to server.
-                """
-            )
-            let metadata = SendableItemMetadata(
-                ocId: itemTemplate.itemIdentifier.rawValue,
-                account: account.ncKitAccount,
-                classFile: NKCommon.TypeClassFile.unknow.rawValue,
-                contentType: itemTemplate.contentType?.preferredMIMEType ?? "",
-                creationDate: itemTemplate.creationDate as? Date ?? Date(),
-                date: Date(),
-                directory: false,
-                e2eEncrypted: false,
-                etag: "",
-                fileId: itemTemplate.itemIdentifier.rawValue,
-                fileName: itemTemplate.filename,
-                fileNameView: itemTemplate.filename,
-                hasPreview: false,
-                iconName: "",
-                mountType: "",
-                ownerId: account.id,
-                ownerDisplayName: "",
-                path: relativePath,
-                serverUrl: parentItemRemotePath,
-                size: itemTemplate.documentSize??.int64Value ?? 0,
-                status: Status.normal.rawValue,
-                downloaded: true,
-                uploaded: false,
-                urlBase: account.serverUrl,
-                user: account.username,
-                userId: account.id
-            )
-            dbManager.addItemMetadata(metadata)
-            let item = Item(
-                metadata: metadata,
-                parentItemIdentifier: parentItemIdentifier,
+            return Item.createIgnored(
+                basedOn: itemTemplate,
+                parentItemRemotePath: parentItemRemotePath,
+                contents: url,
                 account: account,
                 remoteInterface: remoteInterface,
+                progress: progress,
                 dbManager: dbManager
             )
-
-            if #available(macOS 13.0, *) {
-                return (item, NSFileProviderError(.excludedFromSync))
-            } else {
-                return (item, nil)
-            }
         }
 
         let fileNameLocalPath = url?.path ?? ""
